% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_DesSocialData.R
\name{build_DesSocialData}
\alias{build_DesSocialData}
\title{build_DesSocialData}
\usage{
build_DesSocialData(.data)
}
\arguments{
\item{.data}{data built with munge_DSVExcel}
}
\value{
Desorollo Social Aggragted data
}
\description{
build_DesSocialData
}
