% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_DSVData.R
\name{build_DSVData}
\alias{build_DSVData}
\title{build_DSVData}
\usage{
build_DSVData(.path)
}
\arguments{
\item{.path}{path where <Formato Indicadores Desarrollo Social> excel is saved}
}
\value{
Desorollo Social and Voluntariado Aggragted data
}
\description{
build_DSVData
}
