% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_SaludRFData.R
\name{build_SaludRFData}
\alias{build_SaludRFData}
\title{build_SaludRFData}
\usage{
build_SaludRFData(.data)
}
\arguments{
\item{.data}{munged Salud Data Frame}
}
\value{
.data transfomed Salud Data Frame to fit RF layout
}
\description{
build_SaludRFData
}
