% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_VoluntariadoData.R
\name{build_VoluntariadoData}
\alias{build_VoluntariadoData}
\title{build_VoluntariadoData}
\usage{
build_VoluntariadoData(.data)
}
\arguments{
\item{.data}{data built with munge_DSVExcel}
}
\value{
Voluntariado Aggragted data
}
\description{
build_VoluntariadoData
}
