% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_DSVExcel.R
\name{munge_DSVExcel}
\alias{munge_DSVExcel}
\title{munge_DSVExcel}
\usage{
munge_DSVExcel(.path)
}
\arguments{
\item{.path}{path where <Formato Indicadores Desarrollo Social> excel is saved}
}
\value{
munged Fornmato Indicadores Desarollo Social
}
\description{
munge_DSVExcel
}
