% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_IndiceSaludExcel.R
\name{munge_IndiceSaludExcel}
\alias{munge_IndiceSaludExcel}
\title{munge_IndiceSaludExcel}
\usage{
munge_IndiceSaludExcel(.path)
}
\arguments{
\item{.path}{data path where Indice de Salud Report is saved and
each sheet is a Country with the specific standarized layout}
}
\value{
.data a standarized data frame with a long format for Health index
}
\description{
munge_IndiceSaludExcel
}
