% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_SaludExcel.R
\name{munge_SaludExcel}
\alias{munge_SaludExcel}
\title{munge_SaludExcel}
\usage{
munge_SaludExcel(.path)
}
\arguments{
\item{.path}{data path where Reporte FEMSA Indicadores Salud Género is saved and
each sheet is a Country with the specific standarized layout}
}
\value{
.data a standarized data frame with a long format
}
\description{
munge_SaludExcel
}
