% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_IndiceSaludExcel.R
\name{read_IndiceSaludExcel}
\alias{read_IndiceSaludExcel}
\title{read_IndiceSaludExcel}
\usage{
read_IndiceSaludExcel(path, sheet)
}
\arguments{
\item{path}{data path where Indice de Salud Report is saved and
each sheet is a Country with the specific standarized layout}

\item{sheet}{Country}
}
\value{
read Data
}
\description{
read_IndiceSaludExcel
}
