% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SaludExcel.R
\name{read_SaludExcel}
\alias{read_SaludExcel}
\title{read_SaludExcel}
\usage{
read_SaludExcel(path, sheet)
}
\arguments{
\item{path}{data path where Reporte FEMSA Indicadores Salud Género is saved and
each sheet is a Country with the specific standarized layout}

\item{sheet}{Country}
}
\value{
read Data
}
\description{
read_SaludExcel
}
